feat(workstation): make scripted gestures optional drivers (#18457)
