+++ apps/workstation/view/Workspace.mjs
@@ -19,8 +19,6 @@ import WorkspaceDocument        from '../../../src/dashboard/dock/model/Workspac
 import WorkspaceController      from './WorkspaceController.mjs';
 import Operations               from '../../../src/dashboard/dock/model/Operations.mjs';
 import Persistence              from '../../../src/dashboard/dock/model/Persistence.mjs';
-import PreviewContract          from '../../../src/dashboard/dock/model/PreviewContract.mjs';
-import InteractionService       from '../../../src/ai/client/InteractionService.mjs';
 import StateProvider            from '../../../src/state/Provider.mjs';
 import TourRunner               from '../../../src/ai/client/TourRunner.mjs';
 import TransactionManager       from '../../../src/manager/Transaction.mjs';
@@ -245,11 +243,10 @@ class Workspace extends DockWorkspace {
      * @member {Neo.dashboard.dock.interaction.DragAffordances|null} dragAffordances=null
      */
     dragAffordances = null
-    /**
-     * Real-pointer gesture driver for the app-owned journey executors (spec + film replays).
-     * @member {Neo.ai.client.InteractionService|null} interactionService=null
-     */
-    interactionService = null
+    /** @member {Workstation.tour.NativeGestureDriver|null} gestureDriver=null */
+    gestureDriver = null
+    /** @member {Promise|null} gestureDriverPromise=null */
+    gestureDriverPromise = null
     /**
      * @member {Object} paneCache={}
      * @protected
@@ -497,8 +494,6 @@ class Workspace extends DockWorkspace {
             preview   : me.getReference('dock-preview')
         });
 
-        me.interactionService = Neo.create(InteractionService, {});
-
         me.tearOutEmbodiment = createDockVesselEmbodiment({
             resolvePane: itemId => me.paneCache[itemId]
                 ?? (me.dockModel?.items?.[itemId] && me.resolvePane(itemId, me.dockModel.items[itemId])),
@@ -704,7 +699,7 @@ class Workspace extends DockWorkspace {
                 await this.refreshPromise;
                 return this.pulseScaleSparkline()
             case 'cross-zone-showcase':
-                return this.executeCrossZoneShowcaseStep(cue, cue.options)
+                return (await this.getGestureDriver()).executeCrossZoneShowcaseStep(cue, cue.options)
             case 'theme':
                 return this.setWorkspaceTheme(cue.theme)
             default:
@@ -2700,7 +2695,6 @@ class Workspace extends DockWorkspace {
         return rect && {height: rect.height, width: rect.width, x: rect.x, y: rect.y}
     }
 
-
     /**
      * @summary Admits a full popup Workspace and transfers its pane through one Group transaction.
      * @param {Object} document Pure reducer result; source truth is re-read at the queue head.
@@ -2779,7 +2773,6 @@ class Workspace extends DockWorkspace {
      */
     afterNativeOwnerChange(itemId, entry, connection=null, isMerge=false) {
         let me = this;
-
 
         connection && me.registerVesselWorkspaceTarget({
             app     : Neo.apps[connection.windowId],
@@ -2891,7 +2884,6 @@ class Workspace extends DockWorkspace {
         await super.onNativeWindowRelease(data)
     }
 
-
     /**
      * The engine cleared a vessel's ownership records — on its release, its lease running out, or a
      * refused recovery — and brought a committed item home. The park and native-park machines, the
@@ -2908,450 +2900,6 @@ class Workspace extends DockWorkspace {
         me.vesselParkHandlers.onVesselRetired({itemId, retirement: true});
         me.nativeVesselParkHandlers.onVesselRetired({itemId, retirement: true});
         recovered && me.retireVesselWorkspaceTarget(itemId)
-    }
-
-    /**
-     * @summary Creates the film-only cursor that rides one gesture executor's own coordinates.
-     *
-     * CDP-dispatched pointer events move no OS cursor, so an unassisted take reads as
-     * UI-moving-itself. The create shape carries the required `className`, floating-component
-     * mount pair, stable `document.body` parent, and owning `windowId` for every subsequent
-     * style delta.
-     * @param {Number} clientX
-     * @param {Number} clientY
-     * @param {String|Number} [windowId=this.windowId]
-     * @returns {Neo.component.Base}
-     * @protected
-     */
-    createFilmCursorDot(clientX, clientY, windowId=this.windowId) {
-        let me        = this,
-            cursorDot = Neo.create({
-                className    : 'Neo.component.Base',
-                appName      : me.appName,
-                autoInitVnode: true,
-                autoMount    : true,
-                parentId     : 'document.body',
-                windowId,
-                cls          : ['film-cursor'],
-                style        : {
-                    backgroundColor: 'rgba(255, 90, 0, 0.92)',
-                    borderRadius   : '50%',
-                    boxShadow      : '0 0 10px rgba(255, 90, 0, 0.95), 0 0 3px rgba(255, 255, 255, 0.85)',
-                    display        : 'block',
-                    height         : '16px',
-                    left           : `${clientX - 8}px`,
-                    pointerEvents  : 'none',
-                    position       : 'fixed',
-                    top            : `${clientY - 8}px`,
-                    width          : '16px',
-                    zIndex         : 99999
-                }
-            });
-
-        cursorDot.mountedPromise.then(() => {
-            console.log(`[film-cursor] dot mounted in ${windowId} at client (${clientX}, ${clientY})`)
-        });
-
-        return cursorDot
-    }
-
-    /**
-     * @summary Retires one film cursor from component, VDOM, and physical body-node truth.
-     * @param {Neo.component.Base|null} cursorDot
-     * @returns {Promise<Boolean>} True after a live cursor's physical removal is acknowledged.
-     * @protected
-     */
-    async retireFilmCursorDot(cursorDot) {
-        if (!cursorDot || cursorDot.isDestroyed) {
-            return false
-        }
-
-        const removalReceipt = Neo.applyDeltas(cursorDot.windowId, {
-            action: 'removeNode',
-            id    : cursorDot.vdom.id
-        });
-
-        // Retire component truth without dispatching a second physical delta. Awaiting the
-        // explicit receipt keeps cross-window replacement creation behind source-node removal.
-        cursorDot.destroy();
-        await removalReceipt;
-
-        return true
-    }
-
-    /**
-     * @summary Drives the in-window showcase beat through real simulated pointer input.
-     *
-     * One live tab crosses at least two foreign zones and two placement kinds. Each dwell first
-     * seeds the target zone's indicator menu, then aims at the indicator component's OWN computed
-     * geometry; the receipt comes from `DockDragAffordances`' live `activeCandidate.preview`, never
-     * a parallel DOM or pointer inference. Commit compares the resulting document with the exact
-     * preview operation applied to the pre-gesture document. Cancel proves byte-identical document
-     * truth and fully retired overlay/session state.
-     * @param {Object} step
-     * @param {Object[]} step.dwells Ordered `{targetNodeId, placementKind}` dwell requests.
-     * @param {String} step.itemId The live dock item to drag.
-     * @param {String} step.sourceNodeId The tabs node currently holding it.
-     * @param {'commit'|'cancel'} [step.terminal='commit']
-     * @param {Object} [options={}]
-     * @param {Number} [options.dwellDelay=120] Milliseconds each accepted preview stays visible.
-     * @param {Number} [options.moveDelay=16] Milliseconds between pointer samples.
-     * @param {Number} [options.moveSteps=12] Samples per path leg.
-     * @param {Number} [options.safetyMargin=48] Required inset from every window edge.
-     * @param {Boolean} [options.showCursor=false] Film mode: show the shared synthetic cursor.
-     * @returns {Promise<Object>}
-     */
-    async executeCrossZoneShowcaseStep(step, {dwellDelay=120, moveDelay=16, moveSteps=12, safetyMargin=48, showCursor=false}={}) {
-        let me = this,
-            {dwells=[], itemId, sourceNodeId, terminal='commit'} = step || {},
-            document                                 = me.dockModel,
-            sourceNode                               = document?.nodes?.[sourceNodeId],
-            button                                   = null,
-            cursorDot                                = null,
-            lastPoint                                = null,
-            proxyPopupEnabled                        = null,
-            restoreProxyPopupConfig                  = () => {
-                if (!sortZone || sortZone.isDestroyed || proxyPopupEnabled === null) {
-                    return {after: null, before: proxyPopupEnabled, restored: false}
-                }
-
-                sortZone.enableProxyToPopup = proxyPopupEnabled;
-
-                return {
-                    after   : sortZone.enableProxyToPopup,
-                    before  : proxyPopupEnabled,
-                    restored: sortZone.enableProxyToPopup === proxyPopupEnabled
-                }
-            },
-            sortZone                                 = null,
-            tabs                                     = null,
-            windowRect                               = null;
-
-        if (!itemId || sourceNode?.type !== 'tabs' || !sourceNode.items.includes(itemId)) {
-            return {applied: false, errors: ['cross-zone showcase must name a live item held by its source tabs node']}
-        }
-
-        if (!Array.isArray(dwells) || dwells.length < 2
-            || new Set(dwells.map(dwell => dwell?.targetNodeId)).size < 2
-            || new Set(dwells.map(dwell => dwell?.placementKind)).size < 2
-            || dwells.some(dwell => !dwell?.targetNodeId || !dwell?.placementKind || dwell.targetNodeId === sourceNodeId)) {
-            return {applied: false, errors: ['cross-zone showcase requires two distinct foreign zones and two distinct placement kinds']}
-        }
-
-        if (!['commit', 'cancel'].includes(terminal)) {
-            return {applied: false, errors: [`unsupported cross-zone terminal '${terminal}'`]}
-        }
-
-        try {
-            await me.refreshPromise;
-
-            let host          = me.getReference('dock-host'),
-                itemIndex     = sourceNode.items.indexOf(itemId),
-                WindowManager = (await import('../../../src/manager/Window.mjs')).default;
-
-            tabs     = host?.down({dockNodeId: sourceNodeId});
-            sortZone = tabs?.getTabBar()?.sortZone;
-            button   = tabs?.getTabAtIndex(itemIndex);
-
-            let window       = WindowManager.get(button?.windowId),
-                [buttonRect] = button ? await button.getDomRect([button.id], button.windowId) : [];
-
-            if (!button || !sortZone || !buttonRect || !window?.innerRect) {
-                return {applied: false, errors: ['cross-zone gesture surfaces are not ready']}
-            }
-
-            windowRect        = window.innerRect;
-            proxyPopupEnabled = sortZone.enableProxyToPopup;
-            // Tear-out hysteresis is measured against the SOURCE TOOLBAR, not the browser edge:
-            // an ordinary cross-zone path necessarily leaves that strip. Disarm only the popup
-            // conversion branch for this gesture; `finally` restores the live config on every
-            // terminal. The window inset below remains a visual-stage safety rule.
-            sortZone.enableProxyToPopup = false;
-
-            let documentBefore = WorkspaceDocument.clone(document),
-                startX         = buttonRect.x + buttonRect.width / 2,
-                startY         = buttonRect.y + buttonRect.height / 2,
-                directionX     = startX > window.innerRect.width  / 2 ? -1 : 1,
-                directionY     = startY > window.innerRect.height / 2 ? -1 : 1,
-                opt            = (clientX, clientY, buttons) => ({
-                    bubbles   : true,
-                    button    : 0,
-                    buttons,
-                    cancelable: true,
-                    clientX,
-                    clientY,
-                    screenX   : window.innerRect.x + clientX,
-                    screenY   : window.innerRect.y + clientY
-                }),
-                safe           = ({x, y}) => Number.isFinite(x) && Number.isFinite(y)
-                    && x >= safetyMargin && y >= safetyMargin
-                    && x <= window.innerRect.width  - safetyMargin
-                    && y <= window.innerRect.height - safetyMargin,
-                releaseAt      = point => ({
-                    clientX: point?.x ?? startX,
-                    clientY: point?.y ?? startY,
-                    screenX: window.innerRect.x + (point?.x ?? startX),
-                    screenY: window.innerRect.y + (point?.y ?? startY)
-                }),
-                waitUntil      = (predicate, attempts=120) => me.waitFor(predicate, {attempts, delay: 16}),
-                moveTo         = (x, y) => {
-                    if (cursorDot) {
-                        cursorDot.style = {...cursorDot.style, left: `${x - 8}px`, top: `${y - 8}px`}
-                    }
-
-                    return me.interactionService.simulateEvent({events: [{
-                        delay  : moveDelay,
-                        // Once the sort starts, its render can replace the source button DOM identity.
-                        // Native drag sensors are document-global after arming, so keep the synthetic
-                        // stream on the same stable owner instead of addressing a stale source node.
-                        targetId: 'document.body',
-                        type    : 'mousemove',
-                        windowId: button.windowId,
-                        options : opt(x, y, 1)
-                    }]})
-                },
-                walkTo         = async point => {
-                    if (!safe(point)) {
-                        throw new Error(`cross-zone path point violates the ${safetyMargin}px window-edge margin`)
-                    }
-
-                    let from     = lastPoint,
-                        distance = from ? Math.hypot(point.x - from.x, point.y - from.y) : 0,
-                        steps    = distance < 1 ? 1 : Math.max(2, Math.floor(moveSteps));
-
-                    for (let index = 1; index <= steps; index++) {
-                        let ratio = index / steps;
-
-                        await moveTo(
-                            Math.round(from.x + (point.x - from.x) * ratio),
-                            Math.round(from.y + (point.y - from.y) * ratio)
-                        )
-                    }
-
-                    lastPoint = point
-                },
-                overlaysRetired = () => !me.dragAffordances.dragGeometry
-                    && !me.dragAffordances.indicators?.candidateSet
-                    && !me.dragAffordances.indicators?.activeCandidate
-                    && !me.dragAffordances.preview?.dockPreview;
-
-            if (!safe({x: startX, y: startY})) {
-                return {applied: false, errors: [`source tab violates the ${safetyMargin}px window-edge margin`]}
-            }
-
-            let armOne = {x: startX + directionX * 8,  y: startY + directionY * 2},
-                armTwo = {x: startX + directionX * 16, y: startY + directionY * 24};
-
-            if (![armOne, armTwo].every(safe)) {
-                return {applied: false, errors: [`drag-arming path violates the ${safetyMargin}px window-edge margin`]}
-            }
-
-            await me.interactionService.simulateEvent({events: [{
-                targetId: button.id,
-                type    : 'mousedown',
-                windowId: button.windowId,
-                options : opt(startX, startY, 1)
-            }, {
-                delay   : 120,
-                targetId: button.id,
-                type    : 'mousemove',
-                windowId: button.windowId,
-                options : opt(armOne.x, armOne.y, 1)
-            }]});
-
-            lastPoint = armOne;
-
-            if (!await me.waitForTearOutDragArmed(sortZone)) {
-                let cancellation = await me.cancelTearOutGesture(
-                    button,
-                    releaseAt(lastPoint),
-                    {sortZone, targetId: 'document.body'}
-                );
-
-                return {applied: false, errors: ['cross-zone drag did not arm'], proof: {cancellation}}
-            }
-
-            // The first threshold-crossing move can replace the source tab DOM. Only after the
-            // sensor reports its document-global ownership do we advance on the stable carrier.
-            await moveTo(armTwo.x, armTwo.y);
-            lastPoint = armTwo;
-
-            showCursor && (cursorDot = me.createFilmCursorDot(lastPoint.x, lastPoint.y));
-
-            let geometry = await me.dragAffordances.ensureGeometry();
-
-            if (!geometry) {
-                throw new Error('cross-zone geometry did not become measurable')
-            }
-
-            let beats        = [],
-                finalPreview = null;
-
-            for (let [index, dwell] of dwells.entries()) {
-                let zone = geometry.zones.find(entry => entry.nodeId === dwell.targetNodeId);
-
-                if (!zone) {
-                    throw new Error(`cross-zone target '${dwell.targetNodeId}' is not measurable`)
-                }
-
-                let seedPoint = {
-                    x: Math.round(zone.rect.x + zone.rect.width  / 2),
-                    y: Math.round(zone.rect.y + zone.rect.height / 2)
-                };
-
-                await walkTo(seedPoint);
-
-                let candidateSetReady = await waitUntil(() =>
-                    me.dragAffordances.indicators?.candidateSet?.zone?.nodeId === dwell.targetNodeId);
-
-                if (!candidateSetReady) {
-                    throw new Error(
-                        `indicator set did not settle on '${dwell.targetNodeId}' — ` +
-                        `current=${me.dragAffordances.indicators?.candidateSet?.zone?.nodeId ?? 'null'} ` +
-                        `seed=(${seedPoint.x},${seedPoint.y}) zones=${geometry.zones.map(entry => entry.nodeId).join(',')} ` +
-                        `dragProxy=${Boolean(sortZone.dragProxy)} startIndex=${sortZone.startIndex} currentIndex=${sortZone.currentIndex}`
-                    )
-                }
-
-                let indicators = me.dragAffordances.indicators,
-                    candidate  = indicators.candidateSet.cross
-                        .find(entry => entry.preview?.placement?.kind === dwell.placementKind),
-                    indicatorPoint = candidate && indicators.getCandidateHitPoint(candidate.preview?.previewId);
-
-                if (!candidate || !indicatorPoint) {
-                    throw new Error(`indicator hit geometry is unavailable for '${dwell.placementKind}'`)
-                }
-
-                await walkTo(indicatorPoint);
-
-                let candidateActive = await waitUntil(() =>
-                    indicators.activeCandidate?.preview?.previewId === candidate.preview.previewId);
-
-                if (!candidateActive) {
-                    throw new Error(`indicator '${candidate.preview.previewId}' never became active`)
-                }
-
-                dwellDelay > 0 && await me.timeout(dwellDelay);
-
-                // The dwell is a TOCTOU window: the candidate verified active above can be lost
-                // while it elapses — a human-pause dwell can outlive the gesture claim's
-                // arbitration TTL, retiring the candidate mid-gesture — or be preempted by a
-                // competing candidate. Re-verify identity before the read: a loss or swap fails
-                // the step through the receipts channel naming the gate, instead of an
-                // unattributed null-read or a silently adopted wrong preview here.
-                let preview = indicators.activeCandidate?.preview;
-
-                if (preview?.previewId !== candidate.preview.previewId) {
-                    throw new Error(
-                        `active candidate '${candidate.preview.previewId}' lost during the ${dwellDelay}ms dwell — ` +
-                        `gate=dwell-reverify active=${preview?.previewId ?? 'null'} ` +
-                        `dwell=${index + 1}/${dwells.length} target='${dwell.targetNodeId}' placement='${dwell.placementKind}'`
-                    )
-                }
-
-                finalPreview = JSON.parse(JSON.stringify(preview));
-                beats.push({
-                    dwell        : index + 1,
-                    placementKind: preview.placement.kind,
-                    previewId    : preview.previewId,
-                    targetNodeId : preview.target.nodeId
-                })
-            }
-
-            if (terminal === 'cancel') {
-                let cancellation = await me.cancelTearOutGesture(
-                        button,
-                        releaseAt(lastPoint),
-                        {sortZone, targetId: 'document.body'}
-                    ),
-                    retired      = await waitUntil(overlaysRetired),
-                    documentAfter = WorkspaceDocument.clone(me.dockModel),
-                    unchanged     = JSON.stringify(documentAfter) === JSON.stringify(documentBefore),
-                    popupConfig   = restoreProxyPopupConfig();
-
-                return {
-                    applied  : false,
-                    beatLog  : beats,
-                    cancelled: true,
-                    errors   : cancellation.settled && retired && unchanged
-                        ? []
-                        : ['cross-zone cancel left drag, document, or overlay residue'],
-                    proof : {
-                        cancellation,
-                        documentAfter,
-                        documentBefore,
-                        documentsUnchanged: unchanged,
-                        overlaysRetired   : retired,
-                        popupConfig
-                    }
-                }
-            }
-
-            let descriptor     = PreviewContract.previewToOperation(finalPreview),
-                expectedResult = descriptor && Operations.applyOperation(
-                    WorkspaceDocument.clone(documentBefore),
-                    descriptor
-                );
-
-            if (!descriptor || !expectedResult || expectedResult.errors?.length || !expectedResult.document) {
-                throw new Error('final active preview did not resolve to one valid document operation')
-            }
-
-            let expectedDocument   = WorkspaceDocument.clone(expectedResult.document),
-                expectedSerialized = JSON.stringify(expectedDocument);
-
-            await me.interactionService.simulateEvent({events: [{
-                targetId: 'document.body',
-                type    : 'mouseup',
-                windowId: button.windowId,
-                options : opt(lastPoint.x, lastPoint.y, 0)
-            }]});
-
-            let settled = await waitUntil(() => JSON.stringify(me.dockModel) === expectedSerialized);
-
-            settled && await me.refreshPromise;
-
-            let documentAfter          = WorkspaceDocument.clone(me.dockModel),
-                documentMatchesPreview = JSON.stringify(documentAfter) === expectedSerialized,
-                retired                = await waitUntil(overlaysRetired),
-                popupConfig            = restoreProxyPopupConfig(),
-                applied                = settled && documentMatchesPreview && retired && popupConfig.restored;
-
-            return {
-                applied,
-                beatLog: beats,
-                errors : applied ? [] : ['cross-zone commit did not equal the previewed operation or retire its overlays'],
-                proof  : {
-                    descriptor,
-                    documentAfter,
-                    documentBefore,
-                    documentMatchesPreview,
-                    expectedDocument,
-                    finalPreview,
-                    overlaysRetired: retired,
-                    popupConfig
-                }
-            }
-        } catch (error) {
-            let clientX = lastPoint?.x ?? 0,
-                clientY = lastPoint?.y ?? 0;
-
-            button && await me.cancelTearOutGesture(
-                button,
-                {
-                    clientX,
-                    clientY,
-                    screenX: (windowRect?.x ?? 0) + clientX,
-                    screenY: (windowRect?.y ?? 0) + clientY
-                },
-                {sortZone, targetId: 'document.body'}
-            ).catch(() => {});
-
-            return {applied: false, errors: [error?.message || String(error)]}
-        } finally {
-            restoreProxyPopupConfig();
-            await me.retireFilmCursorDot(cursorDot)
-        }
     }
 
     /**
@@ -3444,1005 +2992,23 @@ class Workspace extends DockWorkspace {
     }
 
     /**
-     * @summary Polls one exact cross-window transfer through model adoption and queued projection.
-     * @param {Object} expected
-     * @param {String} expected.sourceWorkspaceId
-     * @param {String} expected.targetWorkspaceId
-     * @param {Object} [options={}]
-     * @param {Number} [options.attempts=240]
-     * @param {Number} [options.delay=16]
-     * @returns {Promise<Object|null>}
-     * @protected
+     * @summary Loads and creates scripted gestures only when a demo or explicit caller needs them.
+     * @returns {Promise<Workstation.tour.NativeGestureDriver>}
      */
-    async waitForCrossWindowTransfer(expected, {attempts=240, delay=16}={}) {
-        let me = this,
-            receipt;
-
-        for (let attempt = 0; attempt <= attempts && !me.isDestroyed; attempt++) {
-            receipt = me.lastCrossWindowTransfer;
-
-            if (
-                receipt?.applied === true && receipt.reconciled === true &&
-                receipt.sourceWorkspaceId === expected?.sourceWorkspaceId &&
-                receipt.targetWorkspaceId === expected?.targetWorkspaceId
-            ) {
-                return receipt
-            }
-
-            attempt < attempts && await me.timeout(delay)
-        }
-
-        return receipt ?? null
+    async getGestureDriver() {
+        const me = this;
+        if (me.isDestroyed) throw Neo.isDestroyed;
+        return me.gestureDriverPromise ??= (async () => {
+            const {default: Driver} = await me.trap(import('../tour/NativeGestureDriver.mjs'));
+            return me.gestureDriver = Neo.create(Driver, {workspace: me})
+        })()
     }
 
-    /**
-     * @summary Scene 3's real-pointer executor: converts a second tear-out while the gesture remains
-     * down, parks that exact OS window over a committed sibling vessel, and releases only after one
-     * semantic + rendered target claim has settled.
-     *
-     * The source stays on the ordinary tab-drag path. It first crosses the source workspace boundary
-     * to acquire its real vessel; subsequent events keep source-local coordinates outside while
-     * moving in global screen space over the target vessel. The landed conversion sensor and
-     * coordinator therefore own the park, preview, arbitration, and atomic transfer. This method
-     * drives and reports those seams; it never calls a reducer or target commit directly.
-     * @param {Object} step
-     * @param {String} step.itemId Incoming pane id.
-     * @param {String} step.sourceNodeId Main-workspace tabs node currently holding the pane.
-     * @param {String} step.targetItemId Existing detached pane whose vessel becomes the target.
-     * @param {Object} [options={}]
-     * @param {Number} [options.attempts=240]
-     * @param {Number} [options.dwellDelay=0] Optional headed-witness hold after target readiness.
-     * @param {Number} [options.moveDelay=16]
-     * @param {Number} [options.moveSteps=4]
-     * @param {Boolean} [options.showCursor=false] Film mode: show the synthetic cursor in whichever
-     * window currently owns the visible leg of the gesture.
-     * @returns {Promise<Object>}
-     */
-    async executeCrossWindowDockStep(
-        step,
-        {attempts=240, dwellDelay=0, moveDelay=16, moveSteps=4, showCursor=false}={}
-    ) {
-        let me                                   = this,
-            {itemId, sourceNodeId, targetItemId} = step || {},
-            targetWorkspaceId                    = Workspace.vesselWorkspaceId(targetItemId),
-            targetState                          = targetWorkspaceId && me.getPopupState(targetWorkspaceId),
-            sourceDocument                       = me.dockModel,
-            sourceNode                           = sourceDocument?.nodes?.[sourceNodeId],
-            button                               = null,
-            cursorDot                            = null,
-            release                              = null,
-            sortZone                             = null;
-
-        if (
-            !itemId || !targetItemId || itemId === targetItemId ||
-            sourceNode?.type !== 'tabs' || !sourceNode.items.includes(itemId)
-        ) {
-            return {applied: false, errors: ['cross-window dock step must name distinct live source and target panes']}
-        }
-        if (
-            !targetState || targetState.committed || targetState.closeRequested ||
-            !me.nativeWindows?.getOwner(me.id, targetItemId)
-        ) {
-            return {applied: false, errors: ['target vessel is not an available first-dock workspace']}
-        }
-
-        let pane = me.paneCache[itemId];
-
-        if (!pane || pane.isDestroyed || !sourceDocument.items?.[itemId]) {
-            return {applied: false, errors: ['incoming pane is not live and owned by the main workspace']}
-        }
-
-        try {
-            await me.refreshPromise;
-            targetState.participationPromise && await targetState.participationPromise;
-
-            let host          = me.getReference('dock-host'),
-                tabs          = host?.down({dockNodeId: sourceNodeId}),
-                itemIndex     = sourceNode.items.indexOf(itemId),
-                WindowManager = (await import('../../../src/manager/Window.mjs')).default;
-
-            button   = tabs?.getTabAtIndex(itemIndex);
-            sortZone = tabs?.getTabBar()?.sortZone;
-
-            let [buttonRect] = button ? await button.getDomRect([button.id], button.windowId) : [],
-                sourceWindow = WindowManager.get(button?.windowId),
-                targetWindow = WindowManager.get(targetState.windowId);
-
-            if (
-                !button || !sortZone || !buttonRect || !sourceWindow?.innerRect ||
-                !targetWindow?.innerRect || !targetState.participation
-            ) {
-                return {applied: false, errors: ['cross-window dock gesture surfaces are not ready']}
-            }
-
-            let startX  = buttonRect.x + buttonRect.width / 2,
-                startY  = buttonRect.y + buttonRect.height / 2,
-                startSX = sourceWindow.innerRect.x + startX,
-                startSY = sourceWindow.innerRect.y + startY,
-                opt     = (clientX, clientY, screenX, screenY, buttons) => ({
-                    bubbles: true, button: 0, buttons, cancelable: true,
-                    clientX, clientY, screenX, screenY
-                }),
-                moveCursor = (clientX, clientY) => {
-                    if (cursorDot) {
-                        cursorDot.style = {
-                            ...cursorDot.style,
-                            left: `${clientX - 8}px`,
-                            top : `${clientY - 8}px`
-                        }
-                    }
-                };
-
-            me.nativeWindows.clearConnection(me.id, itemId);
-            showCursor && (cursorDot = me.createFilmCursorDot(startX, startY, button.windowId));
-
-            await me.interactionService.simulateEvent({events: [{
-                targetId: button.id, type: 'mousedown', windowId: button.windowId,
-                options : opt(startX, startY, startSX, startSY, 1)
-            }, {
-                delay  : 120, targetId: button.id, type: 'mousemove', windowId: button.windowId,
-                options: opt(startX + 8, startY + 2, startSX + 8, startSY + 2, 1)
-            }, {
-                delay  : moveDelay, targetId: button.id, type: 'mousemove', windowId: button.windowId,
-                options: opt(startX + 16, startY + 24, startSX + 16, startSY + 24, 1)
-            }]});
-
-            if (!await me.waitForTearOutDragArmed(sortZone)) {
-                release = {clientX: startX, clientY: startY, screenX: startSX, screenY: startSY};
-
-                let cancellation = await me.cancelTearOutGesture(button, release, {sortZone});
-
-                return {
-                    applied: false,
-                    errors : ['cross-window source drag did not arm'],
-                    proof  : {cancellation}
-                }
-            }
-
-            let boundary = sortZone.boundaryContainerRect,
-                right    = boundary.right  ?? boundary.x + boundary.width,
-                bottom   = boundary.bottom ?? boundary.y + boundary.height,
-                outX     = Math.round(right + 120),
-                outY     = Math.round(bottom + 120);
-
-            for (let index = 1; index <= moveSteps; index++) {
-                let t       = index / moveSteps,
-                    clientX = Math.round(startX + (outX - startX) * t),
-                    clientY = Math.round(startY + (outY - startY) * t);
-
-                moveCursor(clientX, clientY);
-
-                await me.interactionService.simulateEvent({events: [{
-                    delay  : moveDelay, targetId: button.id, type: 'mousemove', windowId: button.windowId,
-                    options: opt(
-                        clientX,
-                        clientY,
-                        sourceWindow.innerRect.x + clientX,
-                        sourceWindow.innerRect.y + clientY,
-                        1
-                    )
-                }]})
-            }
-
-            if (!await me.waitForTearOutVessel(itemId, {attempts})) {
-                release = {
-                    clientX: outX,
-                    clientY: outY,
-                    screenX: sourceWindow.innerRect.x + outX,
-                    screenY: sourceWindow.innerRect.y + outY
-                };
-
-                let cancellation = await me.cancelTearOutGesture(button, release, {sortZone});
-
-                return {
-                    applied: false,
-                    errors : ['cross-window source vessel was not born while the gesture remained down'],
-                    proof  : {cancellation}
-                }
-            }
-
-            let remoteSnapshot;
-
-            if (showCursor) {
-                await me.retireFilmCursorDot(cursorDot);
-                cursorDot = null
-            }
-
-            for (let attempt = 0; attempt <= attempts && !me.isDestroyed; attempt++) {
-                targetWindow = WindowManager.get(targetState.windowId);
-
-                if (!targetWindow?.innerRect) break;
-
-                let targetClientX = Math.round(targetWindow.innerRect.width  / 2) + attempt % 2,
-                    targetClientY = Math.round(targetWindow.innerRect.height / 2),
-                    targetScreenX = Math.round(targetWindow.innerRect.x + targetClientX),
-                    targetScreenY = Math.round(targetWindow.innerRect.y + targetClientY);
-
-                showCursor && !cursorDot &&
-                    (cursorDot = me.createFilmCursorDot(targetClientX, targetClientY, targetState.windowId));
-                moveCursor(targetClientX, targetClientY);
-
-                release = {clientX: outX, clientY: outY, screenX: targetScreenX, screenY: targetScreenY};
-
-                await me.interactionService.simulateEvent({events: [{
-                    delay  : moveDelay, targetId: button.id, type: 'mousemove', windowId: button.windowId,
-                    options: opt(outX + attempt % 2, outY, targetScreenX, targetScreenY, 1)
-                }]});
-
-                let transition = sortZone.vesselConversionSensor?.transitionPromise;
-
-                transition && await transition;
-                remoteSnapshot = me.readCrossWindowGestureSnapshot({
-                    parkedItemId: itemId,
-                    sourceZone  : sortZone,
-                    targetWorkspaceId
-                });
-
-                if (remoteSnapshot.ready) break
-            }
-
-            if (!remoteSnapshot?.ready) {
-                let cancellation = await me.cancelTearOutGesture(button, release, {sortZone});
-
-                return {
-                    applied: false,
-                    errors : ['target vessel did not reach one parked semantic + rendered claim'],
-                    proof  : {cancellation, remoteSnapshot}
-                }
-            }
-
-            dwellDelay > 0 && await me.timeout(dwellDelay);
-
-            me.lastCrossWindowTransfer = null;
-
-            await me.interactionService.simulateEvent({events: [{
-                targetId: button.id, type: 'mouseup', windowId: button.windowId,
-                options : opt(release.clientX, release.clientY, release.screenX, release.screenY, 0)
-            }]});
-
-            let transfer = await me.waitForCrossWindowTransfer({
-                    sourceWorkspaceId: Workspace.MAIN_WORKSPACE_ID,
-                    targetWorkspaceId
-                }, {attempts}),
-                sourceAfter = WorkspaceDocument.clone(me.dockModel),
-                targetAfter = WorkspaceDocument.clone(me.getPopupState(targetWorkspaceId)?.document),
-                retired     = await me.waitForTearOutVesselRetired(itemId, {attempts}),
-                targetItems = targetAfter?.nodes?.[Workspace.vesselTabsNodeId(targetItemId)]?.items || [],
-                sourceOwns  = WorkspaceDocument.findContainingTabsId(sourceAfter, itemId) != null
-                    || WorkspaceDocument.findContainingTabsId(sourceAfter, targetItemId) != null,
-                applied     = transfer?.reconciled === true && retired && !sourceOwns
-                    && targetItems.length === 2
-                    && targetItems[0] === targetItemId
-                    && targetItems[1] === itemId;
-
-            return {
-                applied,
-                errors: applied ? [] : ['cross-window dock did not settle as one A+B target adoption'],
-                proof : {
-                    remoteSnapshot,
-                    sourceDocument     : sourceAfter,
-                    sourceVesselRetired: retired,
-                    targetDocument     : targetAfter,
-                    transfer           : transfer ? WorkspaceDocument.clone(transfer) : null
-                }
-            }
-        } catch (error) {
-            button && await me.cancelTearOutGesture(button, release, {sortZone}).catch(() => {});
-
-            return {applied: false, errors: [error?.message || String(error)]}
-        } finally {
-            await me.retireFilmCursorDot(cursorDot)
-        }
-    }
-
-    /**
-     * @summary Scene 4's real-pointer executor: drags a committed vessel's model-resolved stack
-     * grip home and waits through atomic `transferNode`, main projection, native close request, and
-     * physical topology exit.
-     *
-     * The pointer starts on the actual nested `.neo-dock-stack-handle`, so
-     * {@link Neo.dashboard.dock.interaction.TabSortZone} authors the group payload. The executor never invokes
-     * `transferNode` itself; it withholds mouseup until the main target's one semantic + rendered
-     * claim settles, then observes the resulting receipt through window disconnect.
-     * @param {Object} step
-     * @param {String} step.ownerItemId The pane whose committed vessel owns the stack.
-     * @param {Object} [options={}]
-     * @param {Number} [options.attempts=240]
-     * @param {Number} [options.moveDelay=16]
-     * @param {Boolean} [options.showCursor=false] Film mode: move the synthetic cursor from the
-     * committed vessel into the main window with the physical gesture.
-     * @returns {Promise<Object>}
-     */
-    async executeStackReturnStep(step, {attempts=240, moveDelay=16, showCursor=false}={}) {
-        let me            = this,
-            {ownerItemId} = step || {},
-            workspaceId   = Workspace.vesselWorkspaceId(ownerItemId),
-            state         = workspaceId && me.getPopupState(workspaceId),
-            button        = null,
-            cursorDot     = null,
-            handleId      = null,
-            release       = null,
-            sortZone      = null;
-
-        if (!state?.committed || !state.document || !me.workspaceSet.has(workspaceId)) {
-            return {applied: false, errors: ['stack return requires one committed vessel workspace']}
-        }
-
-        try {
-            await me.refreshPromise;
-            await me.crossWindowParticipationPromise;
-
-            let nodeId        = WorkspaceDocument.resolveStackRoot(state.document),
-                tabsNodeId    = Workspace.vesselTabsNodeId(ownerItemId),
-                tabsNode      = state.document.nodes?.[tabsNodeId],
-                activeItemId  = tabsNode?.activeItemId ?? tabsNode?.items?.[0],
-                activeIndex   = tabsNode?.items?.indexOf(activeItemId) ?? -1,
-                tabs          = state.host?.down({dockNodeId: tabsNodeId}),
-                WindowManager = (await import('../../../src/manager/Window.mjs')).default;
-
-            button   = tabs?.getTabAtIndex(activeIndex);
-            sortZone = tabs?.getTabBar()?.sortZone;
-            handleId = DockLayoutAdapter.stackHandleDomId(activeItemId);
-
-            let [handleRect] = button && handleId
-                    ? await button.getDomRect([handleId], button.windowId)
-                    : [],
-                sourceWindow = WindowManager.get(state.windowId),
-                targetWindow = WindowManager.get(me.windowId);
-
-            if (
-                !nodeId || !button || !sortZone || !handleRect || !sourceWindow?.innerRect ||
-                !targetWindow?.innerRect || !me.crossWindowParticipations.get(Workspace.MAIN_WORKSPACE_ID)
-            ) {
-                return {applied: false, errors: ['whole-stack gesture surfaces are not ready']}
-            }
-
-            let startX  = handleRect.x + handleRect.width / 2,
-                startY  = handleRect.y + handleRect.height / 2,
-                startSX = sourceWindow.innerRect.x + startX,
-                startSY = sourceWindow.innerRect.y + startY,
-                opt     = (clientX, clientY, screenX, screenY, buttons) => ({
-                    bubbles: true, button: 0, buttons, cancelable: true,
-                    clientX, clientY, screenX, screenY
-                }),
-                moveCursor = (clientX, clientY) => {
-                    if (cursorDot) {
-                        cursorDot.style = {
-                            ...cursorDot.style,
-                            left: `${clientX - 8}px`,
-                            top : `${clientY - 8}px`
-                        }
-                    }
-                };
-
-            showCursor && (cursorDot = me.createFilmCursorDot(startX, startY, state.windowId));
-
-            await me.interactionService.simulateEvent({events: [{
-                targetId: handleId, type: 'mousedown', windowId: button.windowId,
-                options : opt(startX, startY, startSX, startSY, 1)
-            }, {
-                delay  : 120, targetId: handleId, type: 'mousemove', windowId: button.windowId,
-                options: opt(startX + 8, startY, startSX + 8, startSY, 1)
-            }, {
-                delay  : moveDelay, targetId: handleId, type: 'mousemove', windowId: button.windowId,
-                options: opt(startX + 24, startY, startSX + 24, startSY, 1)
-            }]});
-
-            let armed = false;
-
-            for (let attempt = 0; attempt <= 120 && !me.isDestroyed; attempt++) {
-                armed = Boolean(sortZone.stackDragActive && sortZone.dragProxy && sortZone.dragCoordinator);
-
-                if (armed) break;
-
-                attempt < 120 && await me.timeout(16)
-            }
-
-            if (!armed) {
-                release = {clientX: startX, clientY: startY, screenX: startSX, screenY: startSY};
-
-                let cancellation = await me.cancelTearOutGesture(button, release, {sortZone, targetId: handleId});
-
-                return {
-                    applied: false,
-                    errors : ['whole-stack source drag did not arm from the rendered grip'],
-                    proof  : {
-                        cancellation,
-                        sourceArm: {
-                            dockGroupNodeId: sortZone.dockGroupNodeId,
-                            dragComponent  : sortZone.dragComponent?.id ?? null,
-                            dragProxyReady : Boolean(sortZone.dragProxy),
-                            stackDragActive: sortZone.stackDragActive === true,
-                            startIndex     : sortZone.startIndex
-                        }
-                    }
-                }
-            }
-
-            let remoteSnapshot;
-
-            if (showCursor) {
-                await me.retireFilmCursorDot(cursorDot);
-                cursorDot = null
-            }
-
-            // The film gesture aims at the semantic return target itself: the indicator
-            // menu's active candidate is selected geometrically, so the synthetic cursor
-            // hovers the stored-home tabs node (window center can lie outside it).
-            let storedHome   = me.tearOutHandlers?.peekPlacement?.(state.itemId)?.tabsNodeId,
-                returnNodeId = me.dockModel.nodes?.[storedHome]?.type === 'tabs'
-                    ? storedHome
-                    : Object.entries(me.dockModel.nodes || {}).find(([, node]) => node.type === 'tabs')?.[0],
-                returnHost   = me.dragAffordances?.host,
-                returnNode   = returnNodeId && returnHost?.down({dockNodeId: returnNodeId}),
-                [returnRect] = returnNode ? await returnHost.getDomRect([returnNode.id], me.windowId) : [];
-
-            for (let attempt = 0; attempt <= attempts && !me.isDestroyed; attempt++) {
-                targetWindow = WindowManager.get(me.windowId);
-
-                if (!targetWindow?.innerRect) break;
-
-                let targetClientX = returnRect
-                        ? Math.round(returnRect.x + returnRect.width  / 2) + attempt % 2
-                        : Math.round(targetWindow.innerRect.width  / 2) + attempt % 2,
-                    targetClientY = returnRect
-                        ? Math.round(returnRect.y + returnRect.height / 2)
-                        : Math.round(targetWindow.innerRect.height / 2),
-                    targetScreenX = Math.round(targetWindow.innerRect.x + targetClientX),
-                    targetScreenY = Math.round(targetWindow.innerRect.y + targetClientY);
-
-                showCursor && !cursorDot &&
-                    (cursorDot = me.createFilmCursorDot(targetClientX, targetClientY, me.windowId));
-                moveCursor(targetClientX, targetClientY);
-
-                release = {
-                    clientX: startX + 32,
-                    clientY: startY,
-                    screenX: targetScreenX,
-                    screenY: targetScreenY
-                };
-
-                await me.interactionService.simulateEvent({events: [{
-                    delay  : moveDelay, targetId: handleId, type: 'mousemove', windowId: button.windowId,
-                    options: opt(
-                        release.clientX + attempt % 2,
-                        release.clientY,
-                        release.screenX,
-                        release.screenY,
-                        1
-                    )
-                }]});
-
-                remoteSnapshot = me.readCrossWindowGestureSnapshot({
-                    sourceZone       : sortZone,
-                    targetWorkspaceId: Workspace.MAIN_WORKSPACE_ID
-                });
-
-                if (remoteSnapshot.ready) break
-            }
-
-            if (!remoteSnapshot?.ready) {
-                let cancellation = await me.cancelTearOutGesture(button, release, {sortZone, targetId: handleId});
-
-                return {
-                    applied: false,
-                    errors : ['main workspace did not reach one semantic + rendered stack-return claim'],
-                    proof  : {cancellation, remoteSnapshot}
-                }
-            }
-
-            let sourceWindowId = state.windowId,
-                sourceItemIds  = [
-                    ...(state.document.nodes?.[nodeId]?.items || Object.keys(state.document.items || {}))
-                ];
-
-            me.lastCrossWindowTransfer = null;
-
-            await me.interactionService.simulateEvent({events: [{
-                targetId: handleId, type: 'mouseup', windowId: button.windowId,
-                options : opt(release.clientX, release.clientY, release.screenX, release.screenY, 0)
-            }]});
-
-            let transfer = await me.waitForCrossWindowTransfer({
-                sourceWorkspaceId: workspaceId,
-                targetWorkspaceId: Workspace.MAIN_WORKSPACE_ID
-            }, {attempts});
-
-            for (let attempt = 0; attempt <= attempts && !me.isDestroyed; attempt++) {
-                if (transfer?.topologyExited === true && !WindowManager.get(sourceWindowId)) break;
-
-                attempt < attempts && await me.timeout(16)
-            }
-
-            let mainAfter        = WorkspaceDocument.clone(me.dockModel),
-                targetNodeId     = remoteSnapshot.preview?.target?.nodeId,
-                returnedItems    = mainAfter.nodes?.[targetNodeId]?.items || [],
-                requiredPhases   = ['documents-adopted', 'main-projected', 'close-dispatched', 'topology-exited'],
-                phaseOrder       = (transfer?.phases || []).filter(phase => requiredPhases.includes(phase)),
-                sourceWindowGone = !WindowManager.get(sourceWindowId),
-                applied          = transfer?.descriptor?.operation === 'transferNode'
-                    && transfer.closeRequested === true
-                    && transfer.topologyExited === true
-                    && JSON.stringify(phaseOrder) === JSON.stringify(requiredPhases)
-                    && sourceWindowGone
-                    && sourceItemIds.every(itemId => returnedItems.includes(itemId));
-
-            return {
-                applied,
-                errors: applied ? [] : ['whole-stack return did not settle through model-before-close topology exit'],
-                proof : {
-                    closeReceipt: me.lastTearOutClose ? WorkspaceDocument.clone(me.lastTearOutClose) : null,
-                    mainDocument: mainAfter,
-                    phaseOrder,
-                    remoteSnapshot,
-                    sourceItemIds,
-                    sourceWindowGone,
-                    sourceWindowId,
-                    transfer    : transfer ? WorkspaceDocument.clone(transfer) : null
-                }
-            }
-        } catch (error) {
-            button && await me.cancelTearOutGesture(button, release, {sortZone, targetId: handleId}).catch(() => {});
-
-            return {applied: false, errors: [error?.message || String(error)]}
-        } finally {
-            await me.retireFilmCursorDot(cursorDot)
-        }
-    }
-
-    /**
-     * @summary The app-owned tear-out journey executor — scene 2's real-pointer drive.
-     *
-     * Arms a tab drag, flings the proxy past the window boundary so
-     * {@link Neo.dashboard.dock.interaction.TabSortZone} fires `dockTearOutExit`, the host opens a `?popout=`
-     * vessel, then — gated on that vessel's ACTUAL birth (its slot binding through
-     * {@link Neo.manager.transaction.NativeLifecycle#onBind}) — survives
-     * deliberate post-birth moves and settles one of three terminals: release while detached
-     * (`dockTearOutTerminal` → the `detachItem` commit + adoption), Escape-cancel (zero-mutation
-     * vessel close), or RE-ENTRY (`reenter` — the drag walks back inside past the reattach
-     * threshold, the vessel retires MID-GESTURE with zero mutation and the in-window proxy
-     * resumes: the film's back-IN morph beat, then cancelled cleanly).
-     *
-     * The proof is OBSERVABLE-ONLY — committed document truth + vessel bookkeeping, never the
-     * machine's internal drag state. Path and pace are tunable for the film takes (D-010):
-     * `curve` bows the pointer path (0 = straight, deterministic either way), `moveSteps` and
-     * `moveDelay` set the sampling density and rhythm.
-     * @param {Object} step
-     * @param {String} step.itemId The dock item to tear out.
-     * @param {String} step.sourceNodeId The tabs node currently holding it.
-     * @param {Object} [options={}]
-     * @param {Number} [options.birthAttempts=180] Vessel-birth poll attempts (16ms each) — film
-     *     pacing widens this: vsync-limited boot can push the popup's shared-heap join past the
-     *     default three-second gate.
-     * @param {Boolean} [options.cancel=false] Escape while detached — the zero-mutation witness.
-     * @param {Number} [options.curve=0] Perpendicular path bow as a fraction of path length.
-     * @param {Number} [options.moveDelay=16] Milliseconds between pointer samples.
-     * @param {Number} [options.moveSteps=4] Pointer samples per path leg.
-     * @param {Number} [options.postBirthMoves=2] Outward moves after birth (survival probe; floored at 2).
-     * @param {Boolean} [options.reenter=false] Walk back inside instead of releasing — the morph witness.
-     * @param {Boolean} [options.showCursor=false] Film mode: ride a visible synthetic cursor dot on
-     *     the logged pointer coordinates — CDP events move no OS cursor, and the camera needs one.
-     * @returns {Promise<Object>}
-     */
-    async executeTearOutStep(step, {birthAttempts=180, cancel=false, curve=0, moveDelay=16, moveSteps=4, postBirthMoves=2, reenter=false, showCursor=false}={}) {
-        let me                     = this,
-            {itemId, sourceNodeId} = step || {},
-            document               = me.dockModel,
-            node                   = document?.nodes?.[sourceNodeId],
-            button                 = null,
-            cursorDot              = null,
-            release                = null;
-
-        if (!itemId || node?.type !== 'tabs' || !node.items.includes(itemId)) {
-            return {applied: false, errors: ['tear-out step must name a live item held by a tabs node']}
-        }
-
-        let pane = me.paneCache[itemId];
-
-        if (!pane || pane.isDestroyed || !document.items?.[itemId]) {
-            return {applied: false, errors: ['source pane is not live and owned by the workspace']}
-        }
-
-        try {
-            // Drain the host-owned projection queue before reading tab chrome.
-            await me.refreshPromise;
-
-            let host          = me.getReference('dock-host'),
-                tabs          = host?.down({dockNodeId: sourceNodeId}),
-                sortZone      = tabs?.getTabBar()?.sortZone,
-                itemIndex     = node.items.indexOf(itemId),
-                WindowManager = (await import('../../../src/manager/Window.mjs')).default;
-
-            button = tabs?.getTabAtIndex(itemIndex);
-
-            let window       = WindowManager.get(button?.windowId),
-                [buttonRect] = button ? await button.getDomRect([button.id], button.windowId) : [];
-
-            if (!button || !sortZone || !buttonRect || !window?.innerRect) {
-                return {applied: false, errors: ['tear-out gesture surfaces are not ready']}
-            }
-
-            // The committed document BEFORE the gesture — the zero-mutation (cancel/reenter) and
-            // detach-commit (terminal) proofs both compare against this snapshot.
-            let documentBefore = WorkspaceDocument.clone(document),
-                catalogBefore  = Object.keys(documentBefore.items);
-
-            let startX  = buttonRect.x + buttonRect.width / 2,
-                startY  = buttonRect.y + buttonRect.height / 2,
-                startSX = window.innerRect.x + startX,
-                startSY = window.innerRect.y + startY,
-                opt     = (clientX, clientY, screenX, screenY, buttons) => ({
-                    bubbles: true, button: 0, buttons, cancelable: true, clientX, clientY, screenX, screenY
-                }),
-                moveTo  = (x, y) => {
-                    // The visible cursor rides the executor's own coordinate log — never a
-                    // second derivation that could disagree with what the gesture actually did.
-                    if (cursorDot) {
-                        cursorDot.style = {...cursorDot.style, left: `${x - 8}px`, top: `${y - 8}px`}
-                    }
-
-                    return me.interactionService.simulateEvent({events: [{
-                        delay  : moveDelay, targetId: button.id, type: 'mousemove', windowId: button.windowId,
-                        options: opt(x, y, window.innerRect.x + x, window.innerRect.y + y, 1)
-                    }]})
-                };
-
-            // A stale record from a prior gesture would false-open the birth gate.
-            me.nativeWindows.clearConnection(me.id, itemId);
-
-            // Phase 1: own the native sensor + cross the LOCAL drag arming threshold (delay+distance).
-            await me.interactionService.simulateEvent({events: [{
-                targetId: button.id, type: 'mousedown', windowId: button.windowId,
-                options : opt(startX, startY, startSX, startSY, 1)
-            }, {
-                delay  : 120, targetId: button.id, type: 'mousemove', windowId: button.windowId,
-                options: opt(startX + 8, startY + 2, startSX + 8, startSY + 2, 1)
-            }, {
-                delay  : 16, targetId: button.id, type: 'mousemove', windowId: button.windowId,
-                options: opt(startX + 16, startY + 24, startSX + 16, startSY + 24, 1)
-            }]});
-
-            // The drag arms ASYNC — gate on proxy + boundary + itemRects before any outward sample
-            // or the exit never fires (arming precedes boundary moves).
-            let armed = await me.waitForTearOutDragArmed(sortZone);
-
-            if (!armed) {
-                let cancellation = await me.cancelTearOutGesture(button, {clientX: startX, clientY: startY, screenX: startSX, screenY: startSY});
-
-                return {applied: false, errors: ['tear-out drag did not arm'], proof: {armed: false, cancellation, documentBefore}}
-            }
-
-            if (showCursor) {
-                cursorDot = me.createFilmCursorDot(startX, startY)
-            }
-
-            // Pre-birth entry sentinels: a curved outward path can transiently re-cover the strip
-            // AFTER the exit fired, retiring the newborn vessel before it ever connects — the
-            // birth-failure diag must carry whether that happened, or the death reads as absence.
-            let preBirthBoundaryEntries = 0,
-                preBirthEntries         = 0,
-                preBirthBoundaryProbe   = () => {preBirthBoundaryEntries++},
-                preBirthEntryProbe      = () => {preBirthEntries++};
-
-            sortZone.on('dragBoundaryEntry', preBirthBoundaryProbe);
-            tabs.on('dockTearOutEntry', preBirthEntryProbe);
-
-            // The tear-out exit fires when the proxy LEAVES `boundaryContainerRect`. Target past
-            // its bottom-right corner, fully outside, so intersectionRatio collapses below the
-            // reattach threshold.
-            let b       = sortZone.boundaryContainerRect,
-                bRight  = b.right  ?? (b.x + b.width),
-                bBottom = b.bottom ?? (b.y + b.height),
-                outX    = Math.round(bRight  + 120),
-                outY    = Math.round(bBottom + 120),
-                outSX   = window.innerRect.x + outX,
-                outSY   = window.innerRect.y + outY;
-
-            release = {clientX: outX, clientY: outY, screenX: outSX, screenY: outSY};
-
-            // The deterministic pointer path: a quadratic bow through a perpendicular control
-            // point (curve=0 degenerates to the straight line). t runs 0→1 outward.
-            let pathPoint = t => {
-                let mx = startX + (outX - startX) / 2 + (outY - startY) * curve,
-                    my = startY + (outY - startY) / 2 - (outX - startX) * curve,
-                    ax = startX + (mx - startX) * t, ay = startY + (my - startY) * t,
-                    bx = mx + (outX - mx) * t,       by = my + (outY - my) * t;
-
-                return {x: Math.round(ax + (bx - ax) * t), y: Math.round(ay + (by - ay) * t)}
-            };
-
-            // Phase 2: PROGRESSIVE outward moves — each further out so intersectionRatio steps
-            // down: dragBoundaryExit → dockTearOutExit → the host acquires the vessel.
-            for (let stepIndex = 1; stepIndex <= moveSteps; stepIndex++) {
-                let p = pathPoint(stepIndex / moveSteps);
-
-                await moveTo(p.x, p.y)
-            }
-
-            // Gate on the vessel's ACTUAL birth: a `?popout=` window binding its reserved slot.
-            let born = await me.waitForTearOutVessel(itemId, {attempts: birthAttempts});
-
-            sortZone.un('dragBoundaryEntry', preBirthBoundaryProbe);
-            tabs.un('dockTearOutEntry', preBirthEntryProbe);
-
-            if (!born) {
-                let diag         = `exitFired=${Boolean(sortZone.isWindowDragging)} vesselOpen=${JSON.stringify(me.lastVesselOpen ?? null)} preBirthBoundaryEntries=${preBirthBoundaryEntries} preBirthEntries=${preBirthEntries} lastRatio=${sortZone.lastIntersectionRatio} boundary=${JSON.stringify(b)} out=(${outX},${outY}) itemRects=${sortZone.itemRects?.length ?? 'null'}`,
-                    cancellation = await me.cancelTearOutGesture(button, release);
-
-                return {
-                    applied: false,
-                    errors : [`tear-out vessel was not born after the boundary exit — ${diag}`],
-                    proof  : {armed: true, born: false, diag, cancellation, documentBefore}
-                }
-            }
-
-            // Post-birth survival probe: deliberate OUTWARD moves must NOT reap the newborn vessel.
-            let probeMoves = Math.max(2, postBirthMoves);
-
-            for (let i = 1; i <= probeMoves; i++) {
-                await moveTo(outX, outY + i * 12)
-            }
-
-            let survivedProbe = Boolean(me.nativeWindows?.getConnection(me.id, itemId));
-
-            if (reenter) {
-                // The morph beat: walk back INSIDE until the PROXY re-enters past the reattach
-                // threshold — dockTearOutEntry retires the vessel MID-GESTURE with zero mutation
-                // and the zone resumes its in-window embodiment. Then end the resumed in-window
-                // drag with a clean cancel; the document stays byte-identical.
-                //
-                // The re-entry target is a point ~35% into the boundary rect, NOT the source
-                // button: the tear-out proxy is pane-sized and the grab corner is unknown, so a
-                // button near a window edge can never recover 60% overlap — the interior point
-                // guarantees the ratio for any grab corner.
-                let vesselWindowId    = me.nativeWindows?.getConnection(me.id, itemId)?.windowId ?? null,
-                    inX               = Math.round(b.x + (b.width  ?? 0) * 0.35),
-                    inY               = Math.round(b.y + (b.height ?? 0) * 0.35),
-                    boundaryEntrySeen = false,
-                    entrySeen         = false,
-                    boundaryProbe     = () => {boundaryEntrySeen = true},
-                    entryProbe        = () => {entrySeen = true};
-
-                sortZone.on('dragBoundaryEntry', boundaryProbe);
-                tabs.on('dockTearOutEntry', entryProbe);
-
-                for (let stepIndex = 1; stepIndex <= moveSteps; stepIndex++) {
-                    let t = stepIndex / moveSteps;
-
-                    await moveTo(
-                        Math.round(outX + (inX - outX) * t),
-                        Math.round(outY + (inY - outY) * t)
-                    )
-                }
-
-                let retired     = await me.waitForTearOutVesselRetired(itemId),
-                    reentryDiag = `boundaryEntrySeen=${boundaryEntrySeen} entrySeen=${entrySeen} isWindowDragging=${Boolean(sortZone.isWindowDragging)} reattachArmed=${Boolean(sortZone.reattachArmed)} lastRatio=${sortZone.lastIntersectionRatio} placeholder=${Boolean(sortZone.dragPlaceholder)} indexMap=${JSON.stringify(sortZone.indexMap)} ownerItems=${sortZone.owner?.items?.length} itemRectsLen=${sortZone.itemRects?.length} activeVessel=${Boolean(me.tearOutHandlers.activeVessel)} connects=${Boolean(me.nativeWindows?.getConnection(me.id, itemId))} staged=${me.tearOutEmbodiment.isStaged(itemId)} boundary=${JSON.stringify(b)} in=(${inX},${inY}) vesselDims=${JSON.stringify(me.tearOutVesselDims)}`;
-
-                sortZone.un('dragBoundaryEntry', boundaryProbe);
-                tabs.un('dockTearOutEntry', entryProbe);
-
-                let
-                    cancellation  = await me.cancelTearOutGesture(button, {clientX: inX, clientY: inY, screenX: window.innerRect.x + inX, screenY: window.innerRect.y + inY}),
-                    documentAfter = WorkspaceDocument.clone(me.dockModel),
-                    windowGone    = !vesselWindowId || !WindowManager.get(vesselWindowId);
-
-                return {
-                    applied  : false,
-                    errors   : retired ? [] : [`vessel did not retire on re-entry — ${reentryDiag}`],
-                    reentered: retired,
-                    proof    : {
-                        born              : true,
-                        survivedProbe,
-                        retired,
-                        windowGone,
-                        cancellation,
-                        documentBefore,
-                        documentAfter,
-                        documentsUnchanged: JSON.stringify(documentBefore) === JSON.stringify(documentAfter),
-                        vesselWindowName  : `tearout-${itemId}`
-                    }
-                }
-            }
-
-            if (cancel) {
-                // Escape while detached → dockTearOutCancel → the host closes its vessel. The
-                // committed document must be byte-identical — the zero-mutation invariant.
-                let cancellation  = await me.cancelTearOutGesture(button, release),
-                    documentAfter = WorkspaceDocument.clone(me.dockModel);
-
-                return {
-                    applied  : false,
-                    cancelled: true,
-                    errors   : [],
-                    proof    : {
-                        born              : true,
-                        survivedProbe,
-                        cancellation,
-                        documentBefore,
-                        documentAfter,
-                        documentsUnchanged: JSON.stringify(documentBefore) === JSON.stringify(documentAfter),
-                        vesselWindowName  : `tearout-${itemId}`
-                    }
-                }
-            }
-
-            // Terminal: release while detached → dockTearOutTerminal → the detachItem commit +
-            // adoption (the vessel owns the pane now).
-            await me.interactionService.simulateEvent({events: [{
-                targetId: button.id, type: 'mouseup', windowId: button.windowId,
-                options : opt(outX, outY, outSX, outSY, 0)
-            }]});
-
-            let committed = await me.waitForTearOutCommit(itemId);
-
-            committed && await me.refreshPromise;
-
-            let
-                documentAfter  = WorkspaceDocument.clone(me.dockModel),
-                absentFromTree = !Object.values(documentAfter.nodes).some(zoneNode => zoneNode.items?.includes(itemId)),
-                keptInCatalog  = Boolean(documentAfter.items?.[itemId]);
-
-            return {
-                applied: committed && absentFromTree && keptInCatalog,
-                errors : committed && absentFromTree && keptInCatalog ? [] : ['detachItem commit did not reach committed document truth'],
-                proof  : {
-                    born              : true,
-                    survivedProbe,
-                    committed,
-                    documentBefore,
-                    documentAfter,
-                    detachCommitted   : absentFromTree && keptInCatalog,
-                    itemAbsentFromTree: absentFromTree,
-                    itemKeptInCatalog : keptInCatalog,
-                    catalogPreserved  : catalogBefore.every(id => Boolean(documentAfter.items?.[id])),
-                    vesselWindowName  : `tearout-${itemId}`
-                }
-            }
-        } catch (error) {
-            button && await me.cancelTearOutGesture(button, release).catch(() => {});
-
-            return {applied: false, errors: [error?.message || String(error)]}
-        } finally {
-            // The synthetic cursor is per-gesture presentation: it never outlives the take's
-            // gesture, and it never enters worker truth (pointer-events:none, no dock document).
-            await me.retireFilmCursorDot(cursorDot)
-        }
-    }
-
-    /**
-     * Polls until the base drag has ARMED — a live proxy AND the async main-thread
-     * `boundaryContainerRect` AND measured `itemRects` are all present: exactly the facts
-     * the sort zone needs before it will sample a boundary exit.
-     * @param {Neo.draggable.container.SortZone} sortZone
-     * @param {Object} [options={}]
-     * @param {Number} [options.attempts=120]
-     * @param {Number} [options.delay=16]
-     * @returns {Promise<Boolean>}
-     * @protected
-     */
-    async waitForTearOutDragArmed(sortZone, {attempts=120, delay=16}={}) {
-        let me    = this,
-            armed = () => Boolean(sortZone?.dragProxy && sortZone?.boundaryContainerRect && sortZone?.itemRects);
-
-        return me.waitFor(armed, {attempts, delay})
-    }
-
-    /**
-     * Gates on the tear-out vessel's ACTUAL birth: the `?popout=<itemId>` window binding its
-     * reserved slot ({@link Neo.manager.transaction.NativeLifecycle#onBind}). Polls that observable
-     * rather than any internal drag flag.
-     * @param {String} itemId
-     * @param {Object} [options={}]
-     * @param {Number} [options.attempts=180]
-     * @param {Number} [options.delay=16]
-     * @returns {Promise<Boolean>}
-     * @protected
-     */
-    async waitForTearOutVessel(itemId, {attempts=180, delay=16}={}) {
-        let me = this;
-
-        return me.waitFor(() => Boolean(
-            me.nativeWindows?.getConnection(me.id, itemId) || me.nativeWindows?.getOwner(me.id, itemId)
-        ), {attempts, delay})
-    }
-
-    /**
-     * Gates on a re-entry retirement fully settling: no connect entry, no pending admission,
-     * no staged embodiment, and the tear-out machine's slot cleared.
-     * @param {String} itemId
-     * @param {Object} [options={}]
-     * @param {Number} [options.attempts=180]
-     * @param {Number} [options.delay=16]
-     * @returns {Promise<Boolean>}
-     * @protected
-     */
-    async waitForTearOutVesselRetired(itemId, {attempts=180, delay=16}={}) {
-        let me      = this,
-            retired = () => Boolean(
-                !me.nativeWindows?.getConnection(me.id, itemId) && !Boolean(me.nativeWindows?.getAdmission(me.id, itemId)) &&
-                !me.tearOutEmbodiment.isStaged(itemId) && !me.tearOutHandlers.activeVessel
-            );
-
-        return me.waitFor(retired, {attempts, delay})
-    }
-
-    /**
-     * Gates on the committed detach reaching document truth: the item leaves every node's
-     * `items` (the vessel owns it) while the catalog entry stays.
-     * @param {String} itemId
-     * @param {Object} [options={}]
-     * @param {Number} [options.attempts=180]
-     * @param {Number} [options.delay=16]
-     * @returns {Promise<Boolean>}
-     * @protected
-     */
-    async waitForTearOutCommit(itemId, {attempts=180, delay=16}={}) {
-        let me       = this,
-            detached = () => {
-                let document = me.dockModel;
-
-                return !Object.values(document.nodes).some(zoneNode => zoneNode.items?.includes(itemId))
-                    && Boolean(document.items?.[itemId])
-            };
-
-        return me.waitFor(detached, {attempts, delay})
-    }
-
-    /**
-     * Cancels a live tear-out gesture: an Escape keydown (the drag-cancel signal the sort zone
-     * consumes) followed by a settling mouseup.
-     * @param {Neo.component.Base} button The dragged tab button.
-     * @param {Object} release `{clientX, clientY, screenX, screenY}` the release point.
-     * @param {Object} [options={}]
-     * @param {Neo.draggable.container.SortZone|null} [options.sortZone=null] Zone whose idle facts gate settlement.
-     * @param {String} [options.targetId=button.id] Stable DOM owner for both terminal events.
-     * @returns {Promise<Object>}
-     * @protected
-     */
-    async cancelTearOutGesture(button, release, {sortZone=null, targetId=button?.id}={}) {
-        let me = this;
-
-        if (!button || !targetId) return {escapeDispatched: false, releaseDispatched: false, settled: false};
-
-        let {clientX=0, clientY=0, screenX=0, screenY=0} = release || {},
-            escapeDispatched = await me.interactionService.dispatch({
-                id      : targetId,
-                type    : 'keydown',
-                windowId: button.windowId,
-                options : {bubbles: true, cancelable: true, code: 'Escape', key: 'Escape'}
-            }),
-            releaseDispatched = await me.interactionService.dispatch({
-                id      : targetId,
-                type    : 'mouseup',
-                windowId: button.windowId,
-                options : {bubbles: true, button: 0, buttons: 0, cancelable: true, clientX, clientY, screenX, screenY}
-            });
-
-        if (!sortZone) {
-            return {escapeDispatched, releaseDispatched, settled: true}
-        }
-
-        for (let attempt = 0; attempt <= 60 && !me.isDestroyed; attempt++) {
-            let settled = sortZone.owner?.cls?.includes?.('neo-is-dragging') !== true
-                && sortZone.dragEndActive !== true
-                && sortZone.data == null
-                && !sortZone.dragPlaceholder
-                && !sortZone.dragProxy;
-
-            if (settled) return {escapeDispatched, releaseDispatched, settled: true};
-
-            attempt < 60 && await me.timeout(16)
-        }
-
-        return {escapeDispatched, releaseDispatched, settled: false}
-    }
-
-    /**
-     * @summary Tears down the workspace-owned producer, topology library, tour seams, and panes.
-     * @param {...*} args
-     */
+    /** @summary Tears down the workspace and its optional gesture driver. @param {...*} args */
     destroy(...args) {
         let me = this;
+
+        me.gestureDriver?.destroy();
 
         if (me.#feedIntervalId !== null) {
             clearInterval(me.#feedIntervalId);
@@ -4458,7 +3024,6 @@ class Workspace extends DockWorkspace {
         me.workspaceSet?.destroy();
         me.topologyLibrary?.destroy();
         me.dragAffordances?.destroy();
-        me.interactionService?.destroy();
         me.vesselParkHandlers?.destroy();
         me.nativeVesselParkHandlers?.destroy();
         me.vesselProxyEmbodiment?.destroy();
