+++ buildScripts/util/file-size-baseline.json
@@ -1,5 +1,5 @@
 {
-    "apps/workstation/view/Workspace.mjs": 2670,
+    "apps/workstation/view/Workspace.mjs": 1689,
     "examples/dashboard/crossWindow/DemoBWorkspace.mjs": 2612,
     "src/dashboard/dock/Workspace.mjs": 1035,
     "src/main/addon/DragDrop.mjs": 1267
